+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+<!-- ## [3.0.0] - 2023.11.19 -->
+
 ### Added
 
-- simple grid_world environment (WIP)
+- 2D GridWorld environment 
 
 ## [2.1.0] - 2023.10.24